requirements.txt
Flask
Werkzeug
python-dotenv
langchain
langchain-openai
langchain-community
openai
faiss-cpu
PyPDF2
pandas
tiktoken
flask-cors